tokenization-benchmark/pyproject.toml
[project]
name = "tokenization-benchmark"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.85",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "nvitop>=1.4.2",
    "pandas>=2.2.3",
    "py3dmol>=2.4.2",
    "seaborn>=0.13.2",
]
